Update 02-planning-ahead.Rmd
Hi, 
git hub beginner here, feel free to correct if it's not the good way to propose corrections. 
Found 1 small spelling error and thought it would be a nice way to begin with git_hub interactions on other people project. 
Thanks a lot for the book, it's an amazing helper for me which was never taught how to build apps (phd candidate).